Seems that file size just got longer than designed
https://github.com/rubenlagus/TelegramBots/issues/1088

+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/meta/api/objects/Document.java
@@ -45,5 +45,5 @@ public class Document implements BotApiObject {
     @JsonProperty(MIMETYPE_FIELD)
     private String mimeType; ///< Optional. Mime type of a file as defined by sender
     @JsonProperty(FILESIZE_FIELD)
-    private Integer fileSize; ///< Optional. File size
+    private Long fileSize; ///< Optional. File size
 }
